refactor: 拆分行动的AppSettingsActionSettings文件
