+++ ClassIsland/Models/Actions/CurrentComponentConfigActionSettings.cs
@@ -1,0 +1,17 @@
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+namespace ClassIsland.Models.Actions;
+
+public class CurrentComponentConfigActionSettings : ObservableRecipient
+{
+    string _value = "Default";
+    public string Value
+    {
+        get => _value;
+        set
+        {
+            if (value == _value) return;
+            _value = value;
+            OnPropertyChanged();
+        }
+    }
+}
