+++ ClassIsland/Models/Actions/WindowDockingOffsetXActionSettings.cs
@@ -1,0 +1,17 @@
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+namespace ClassIsland.Models.Actions;
+
+public class WindowDockingOffsetXActionSettings : ObservableRecipient
+{
+    int _value = 0;
+    public int Value
+    {
+        get => _value;
+        set
+        {
+            if (value == _value) return;
+            _value = value;
+            OnPropertyChanged();
+        }
+    }
+}
